.claude/skills/generate-audio/requirements.txt
google-cloud-texttospeech
google-cloud-speech
python-dotenv
